Made project into gradle library
