--- README.md
@@ -1,2 +1,0 @@
-# UDP_TransferNT
-Testing UDP Transfer Protocols, (Possible switch from TCP in CJ-Vision https://github.com/wml-frc/CJ-Vision/)
