--- build.gradle
@@ -1,43 +1,0 @@
-plugins {
-	id "cpp"
-}
-
-task run(type:Exec) {
-	dependsOn(build)
-	workingDir './build/exe/main'
-	commandLine './main'
-}
-
-model {
-	components {
-		main(NativeExecutableSpec) {
-			
-
-			// Sources
-			sources.cpp {   
-				source {
-					srcDirs = ['src/main/cpp']
-					includes = ['**/*.cpp', '**/*.cc', '**/*.c']
-				}
-
-				exportedHeaders {
-					srcDirs = ['src/main/include']
-				}
-				
-			}
-
-			// Binaries
-			binaries.all {
-				if (project.hasProperty("client")) {
-					cppCompiler.define 'CLIENT_RUN'
-				} 
-				
-				if (project.hasProperty("server")) {
-					cppCompiler.define 'SERVER_RUN'
-				}
-
-				linker.args << '-pthread'
-			}
-		}
-	}
-}
