+++ lib/build.gradle
@@ -1,0 +1,22 @@
+plugins {
+	// Apply the cpp-library plugin to add support for building C++ libraries
+	id 'cpp-library'
+
+	// Apply the cpp-unit-test plugin to add support for building and running C++ test executables
+	id 'cpp-unit-test'
+
+}
+
+library {
+	linkage = [Linkage.SHARED]
+
+	targetMachines = [
+		machines.linux.x86_64,
+		machines.windows.x86, machines.windows.x86_64,
+		machines.macOS.x86_64
+	]
+}
+
+tasks.withType(LinkExecutable).configureEach {
+	linkerArgs.add '-pthread'
+}
