+++ lib/src/test/cpp/hello_test.cpp
@@ -1,0 +1,48 @@
+/*
+ * This C++ source file was generated by the Gradle 'init' task.
+ */
+
+#include "UDP_TransferNT.h"
+#include <cassert>
+
+int main() {
+	// assert(greeter.greeting().compare("Hello, World!") == 0);
+
+	#ifdef __linux__
+	std::cout << "Test Start" << std::endl;
+	UDP_TransferNT::DataPacket dpSend, dpRecv;
+	
+	dpSend.setBooleans(1, true);
+	dpSend.setCharacters(10, 't');
+	dpSend.setDecimals(0, 0.111);
+	dpSend.setIntegers(2, 12);
+
+	UDP_TransferNT::Client client(true);
+	UDP_TransferNT::Server server(true);
+
+	client.init();
+	server.init();
+
+	client.registerSend(&dpSend);
+	server.registerRecv(&dpRecv);
+
+	float value = 0;
+
+	while(true) {
+		std::cout << "Registered client sender, and server receiver" << std::endl;
+
+		std::cout << "Booleans index 1: " << dpRecv.getBooleans(1) << std::endl;
+		std::cout << "Characters index 10: " << dpRecv.getCharacters(10) << std::endl;
+		std::cout << "Decimals index 0: " << dpRecv.getDecimals(0) << std::endl;
+		std::cout << "Integers index 2: " << dpRecv.getBooleans(2) << std::endl;
+
+		value += 0.001;
+		dpSend.setDecimals(0, value);
+		sleep(1);
+		system("clear");
+	}
+
+	#else
+	#endif
+	return 0;
+}
