+++ settings.gradle
@@ -8,3 +8,4 @@
  */
 
 rootProject.name = 'UDP_TransferNT'
+include('lib')
