--- src/main/cpp/Client.cpp
@@ -1,17 +1,0 @@
-#include "Client.h"
-
-void Client::init_thread() {
-	this->initNetwork();
-}
-
-void Client::init() {
-		std::cout << "Main Socket Port: " << *this->getSocket()->getPort() << std::endl;
-
-	if (_thread) {
-		std::cout << "Threaded start enabled" << std::endl;
-		std::thread init_t(&Client::init_thread, this);
-		init_t.detach();
-	} else {
-		this->initNetwork();
-	}
-}
