--- src/main/include/main.h
@@ -1,8 +1,0 @@
-#ifndef MAIN_H
-#define MAIN_H
-
-#include "nt_headers.h"
-#include "Client.h"
-#include "Server.h"
-
-#endif
